[VirtualKeyBoard] Catch up fixes
- Harden against missing images

This change hardens the VirtualKeyBoard code against a crash if the "vkey_sel.png" image is not found. The code should now be stable if any or all the VirtualKeyBoard images are missing. The display will be incorrect but the code will no longer crash.

- Correct Estonian keymap

- Fix typo in Polish layout.  Capital and lowercase ł letters were swapped.

- Change the name of the function called when the selected button is activated.

This reverts an inappropriate change made by OpenPLi.