[NumericalTextInput.py] Enforce character restrictions
This code was accidentally omitted from the initial commit.

The key mapping options naturally restrict character input to the listed characters in the mapping lists, this restriction is not enforced for external keyboard input!  This code restoration ensures that only characters that are permitted can be accepted.

The issue could best be seen if HEX input is required.  Using the SMS data entry mode only the characters 0-9, A-F and a-f (as appropriate) can be entered.  Without this code change an external keyboard could enter any character on the keyboard.  With this code change only the permitted 0-9, A-F and a-f (as appropriate) will be accepted.