[config.py] Fix extra_args management
There is a critical flaw in the ConfigElement notifier code with particular reference to the handling of the "extra_args" parameter. The current code tries to use a hash of the notifier method as the key to a dictionary. This is illegal! The errors have been hidden by try / except blocks rather than trying to understand or fix the underlying problem.  The effect of the problem was that attempts to access the extra_args would fail with "TypeError: unhashable type: 'xyz'" and the version of the callback without the "extra_args" would be used instead.  (This causes the "extra_args" information to be lost.)

To address this issue two new methods have been created "addNotifierWithArgs()" and "removeNotifierWithArgs()" that will correctly associate the args with the notifier to which they apply.  For code compatibility any calls to "addNotifier()" with an "extra_args" parameter will be passed to "addNotifierWithArgs()" for correct processing.  A warning of this redirection will be logged so that coders can update their code to use the working mechanisms.  The protective try / except protection has been removed as the underlying issue has been resolved.

While correcting the "extra_args" I have also taken the opportunity to harmonise the internal storage names of the notifiers.  "notifiers_final" is now called "notifiersFinal" to better match the variable naming conventions used elsewhere in this code.  I deprecated the old name as it did not follow the style for all the other notifier names. Consistency and predictability makes it easier to remember variable names and spot mistakes.

Thanks to IanB for assisting address this issue.