[VirtualKeyBoard.py] Replace 'ALL' button with 'DEL'
This change removes the 'ALL' button from the keyboard grid as it now serves no purpose.  The functionality remains in the code should it be desired or useful in the future.  Its position on the keyboard layout is now used by the new 'DEL' button.  This change allows all the locale keyboards to return to their previous layouts and have the 'DEL' button in a consistent position.