Adjust default display skins
This change adjusts the default display skins to work with the new "skin.py" and "SkinSelector.py".  With the new code all skins, including the default skins, are structured as a normal skin with the xml files and the preview images located at the top level directory of the skin.