[mytest.py] Fix boot loop / crash
Some receivers have an issue with the new Timezones.py code causing a crash / boot loop in StbHardware.py.  The problem is that on some receivers the UsageConfig.py is already initialised and on others it isn't.  This fix is to move the Timezones.py initialisation until after the UsageConfig.py is formally initialised.