[InfoBarGenerics.py] Support centralised menu path
The clearScreenPath() method resets the menu path history when the user returns to the main screen.

The current event setTitle() is used to send the current show to the front panel but it should not be used in the  menu history.  To stop the event name from being harvested the new "addToPathList=False" argument is used.