[GUISkin.py] Centralise the menu path code
This change makes the code that calculates and displays the menu path a common piece of code that all Enigma2 code can use without specifically needing to manage the menu path internally to selected modules.

This change also adds a new optional argument to the setTitle() method that can suppress the path display for the current screen.  This may be useful on screens like MessageBox where a full path display may be undesirable.

This change adds a new screen widget called "ScreenPath" that can be used by skinners to share the feature with other images that also have the feature but use different screen widget names. The code automatically maps the new menu path logic onto the existing "menu_path_compressed" widget.