Centralise the menu path code (#183)
* [GUISkin.py] Delete file move content to Screen.py

This file was only ever used by Screen.py.  For efficiency and to simplify code tracing and management this file has been deleted and its contents integrated into Screen.py.

* [Screen.py] Merge in old GUISkin code

- Merge in GUISkin.py code.
- Use Python 3 friendly print syntax.
- Organise code methods.
- camelCase variables.
- Add debugging prints to enhance testing and evaluation of the code.
- Add comments with a "DEBUG:" label to pose questions or suggestions to be considered during the review.

* [mytest.py] Adjust Screen.py variable reference

Change the variable name "Screen.global_screen" to "Screen.globalScreen" to match the updated name in Screen.py.

* [Menu.py] Remove old menu path code

This change removed the old menu path code as this functionality is now performed by the setTitle() method in Screen.py.

* [Setup.py] Remove old menu path code

This change removed the old menu path code as this functionality is now performed by the setTitle() method in Screen.py.