[Setup.py] Add Setup menu image facility
Add a way to have images associated with each Setup menu screen.  These images are defined in the skin via the "<setups>" block previously added during the "skin.py" refactor.

If a specific Setup menu does not have an image defined then a generic "default" image, if available, will be used.  If no images are defined then no "menuimage" widget is created.  For consistency if Setup screen menu images are desired then it is appropriate to ensure that a "default" image is defined and available.

Any skin can have a "<setups>" block contains lines of "<setup>" tags that each contain a "key" attribute is the same as the setup.xml "key" attribute and a "name" attribute is the image file to be used for that Setup screen.