[Screen.py] Add option to suppress screen path
Add a new calling argument "showPath" to the setTitle() method to suppress the setting of the title to include the full screen path.  The default is to follow the current behaviour of modifying the title to include the screen path if the show menu path option is set to "Large".