[Screen.py] Add new "ScreenSummary" class
This class is designed as a more consistent application of a summary (display) screen to be used as a base class for all summary screens.  This class addresses the issue in the "SimpleSummary" that relies on the parent screen to provide a screen title.  The issue with this is that when the show path option is set to "Large" the the main screen displays the screen title with the full menu path history.  This is *not* desirable on the summary screen.

This class is intended to replace the "SimpleSummary" class.