[Setup.py] Switch to new "ScreenSummary" class
Make the summary screen based on the "SummaryScreen" class rather than the GUI "Screen" class.

Note that this will trigger a need to update the "SetupSummary" screen in display skins.  Also note that the widgets have new names.  The "SetupTitle" widget is now "Title", "SetupEntry" is now "entry" and "SetupValue" is now "value".