[Screen.py] Add facility for mandatory widgets
This option works in conjunction with a change in Skin.py to allow code to check if a screen in the skin is compatible with the code.  If the list of widgets considered mandatory are not present then the current screen will be rejected and the next option will be tested.  It is assumed that if all skin options are rejected that the embedded screen will be acceptable.